+++ packages/spacecat-shared-data-access/src/models/fix-entity/fix-entity.collection.js
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 Adobe. All rights reserved.
+ * Copyright 2025 Adobe. All rights reserved.
  * This file is licensed to you under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License. You may obtain a copy
  * of the License at http://www.apache.org/licenses/LICENSE-2.0
@@ -9,19 +9,18 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
+import BaseCollection from '../base/base.collection.js';
 
-export type * from './audit';
-export type * from './configuration';
-export type * from './base';
-export type * from './fix-entity';
-export type * from './experiment';
-export type * from './import-job';
-export type * from './import-url';
-export type * from './key-event';
-export type * from './latest-audit';
-export type * from './opportunity';
-export type * from './organization';
-export type * from './site';
-export type * from './site-candidate';
-export type * from './site-top-page';
-export type * from './suggestion';
+/**
+ * SiteCandidateCollection - A collection class responsible for managing FixEntities.
+ * Extends the BaseCollection to provide specific methods for interacting with
+ * FixEntity records.
+ *
+ * @class FixEntityCollection
+ * @extends BaseCollection
+ */
+class FixEntityCollection extends BaseCollection {
+  // add custom methods here
+}
+
+export default FixEntityCollection;